Fix some tests and get rid of many warnings.
	* mln/value/int_u_sat.hh (int_u_sat<n>::int_u_sat(int))
	(int_u_sat<n>::operator=(int))
	* tests/accu/max_h.cc,
	* tests/accu/median.cc,
	* tests/accu/min_h.cc,
	* tests/labeling/level.cc,
	* tests/line_piter.cc,
	* tests/linear_log.cc,
	* tests/win/disk2d.cc,
	* tests/win/octagon2d.cc: Pacify the compiler w.r.t. comparisons
	between signed and unsigned values.
	* tests/border/fill.cc: Likewise.
	(check): Fix the return type of this function.
	* tests/win/diag2d.cc: Use English in messages, not French.
	* tests/value/graylevel.cc: Test the division of a gl8 value by an
	int_s8 value.
	* tests/line_convolve.cc: Adjust the call to mln::io::load to
	match the current interface.
	* tests/value/float01_bis.cc,
	* tests/sparse_image.cc,
	* tests/rle_image.cc
	Disable these tests.
	* tests/Makefile.am (check_PROGRAMS): Remove chamfer,
	fun_x2x_composed, fun_x2x_rotation, fun_x2x_translation, h_vec and
	run_pset.
	(chamfer_SOURCES, fun_x2x_composed_SOURCES)
	(fun_x2x_rotation_SOURCES, fun_x2x_translation_SOURCES)
	(run_pset_SOURCES): Remove.
	(h_vec_SOURCES): Move...
	* tests/core/Makefile.am: ...here.
	(check_PROGRAMS): Add h_vec.
	* tests/value/builtin.cc: Move...
	* tests/value/builtin/builtin.cc: ...here.
	Pacify the compiler w.r.t. unused variables.
	* tests/value/Makefile.am (check_PROGRAMS): Remove builtin.
	(builtin_SOURCES): Move...
	* tests/value/builtin/Makefile.am: ...here.
	(check_PROGRAMS, TESTS): New.


git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@1653 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/tests/win/disk2d.cc
@@ -1,4 +1,4 @@
-// Copyright (C) 2007 EPITA Research and Development Laboratory
+// Copyright (C) 2007, 2008 EPITA Research and Development Laboratory (LRDE)
 //
 // This file is part of the Olena Library.  This library is free
 // software; you can redistribute it and/or modify it under the terms
@@ -53,7 +53,8 @@ int main()
     {
       mln_assertion(((abs(x) <= 27) && (abs(y) <= 27)) ||
 		    !disk.has(dpoint2d(y, x)));
-      mln_assertion(((x * x + y * y) <= l2) == (disk.has(dpoint2d(y, x))) ||
+      mln_assertion(((x * x + y * y) <= static_cast<int>(l2)) ==
+		    (disk.has(dpoint2d(y, x))) ||
 		    abs(x) > 27 ||
 		    abs(y) > 27);
     }
